requirements.txt



# faiss-cpu==1.10.0
# utils==1.0.2


# pydantic==2.11.3
# torch==2.6.0
# transformers==4.51.3
# openai==1.75.0
# pandas==2.2.2
# numpy==2.0.2
# ipython==7.34.0

# ipywidgets==7.7.1
# requests==2.32.3
# aiohttp==3.11.15
# python-dateutil==2.8.2
# matplotlib==3.10.0
# plotly==5.24.1
# pyyaml==6.0.2

# embedchain==0.1.128   


# # core AI stack
# crewai==0.114.0
# crewai-tools==0.40.1

# # vector store + RAG
# #chromadb>=0.5.23,<0.6.0

# # embeddings + LLM
# langchain>=0.0.250
langchain-huggingface>=0.1.2
# langchain-openai>=0.0.28

# # HF embeddings & transformers
sentence-transformers>=2.2.2
# #transformers==4.20.3
tokenizers==0.21.1
# accelerate>=0.18.0

# # Torch if you need GPU/CPU inference
# torch>=1.13.1
# torchvision>=0.14.1
# torchaudio>=0.13.1

# # other utilities

#tokenizers==0.21.1
#tokenizers
sentence-transformers>=2.2.2

huggingface-hub>=0.12.1

python-dotenv==1.1.0
langchain-community==0.3.22
langchain_openai 
faiss-cpu==1.10.0
utils==1.0.2

sentence_transformers==3.4.1

pydantic==2.11.3
langchain==0.3.24
torch==2.6.0
transformers==4.51.3
openai==1.75.0
pandas==2.2.2
numpy==2.0.2
ipython==7.34.0

ipywidgets==7.7.1
requests==2.32.3
aiohttp==3.11.15
python-dateutil==2.8.2
matplotlib==3.10.0
plotly==5.24.1
pyyaml==6.0.2

# crewai>=0.105.0
# crewai_tools>=0.40.1
# embedchain>=0.1.128 
crewai
crewai_tools
embedchain

gdown
streamlit
folium 
streamlit-folium